+++ Examples/Examples/CounterApp/Models/Counter.swift
@@ -1,5 +1,5 @@
 //
-//  GifList+Transitions.swift
+//  Counter.swift
 //  Examples
 //
 //  Created by Thibault Wittemberg on 2021-02-28.