+++ Examples/ExamplesTests/GiphyApp/GifList+TransitionsTests.swift
@@ -1,0 +1,14 @@
+//
+//  GiphyApp+TransitionsTests.swift
+//  ExamplesTests
+//
+//  Created by Thibault Wittemberg on 2021-02-28.
+//
+
+@testable import Examples
+import Feedbacks
+import FeedbacksTest
+import XCTest
+
+final class GiphyApp_TransitionsTests: XCTestCase {
+}