+++ Sources/Feedbacks/StateMachine/Transitions.swift
@@ -5,66 +5,68 @@
 //  Created by Thibault Wittemberg on 2020-12-23.
 //
 
-/// Represents a series of Transitions to form a State Machine.
-public struct Transitions: TransitionsDefinition {
-    let transitions: [TransitionsDefinition]
+/// Represents a series of Transitions that drive a State Machine.
+public struct Transitions {
+    let transitions: [From]
+
+    /// the reducer computed from the state machine's transitions
+    public let reducer: (State, Event) -> State
 
     /// - Parameter transitions: the individual transitions composing the state machine
-    /// `Transitions {`
-    ///     `Transition(from: LoadingState.self, on: LoadedEvent.self, then: LoadedState())`
-    ///     `Transition(from: LoadingState.self, on: ErrorEvent.self, then: ErrorState())`
-    /// `}`
-    public init(@TransitionsDefinitionsBuilder _ transitions: () -> [TransitionsDefinition]) {
-        self.transitions = transitions()
+    /// Transitions {
+    ///     From(Loading.self) { state in
+    ///       On(LoadingHasComplete.self) { event in
+    ///         Loaded()
+    ///       }
+    ///     }
+    ///     From(Loading.self) { state in
+    ///       On(LoadingHasFailed.self) { event in
+    ///         Failed()
+    ///       }
+    ///     }
+    /// }
+    public init(@ArrayBuilder<From> _ transitions: () -> [From]) {
+        self.init(transitions: transitions())
     }
 
-    init(transitions: [TransitionsDefinition]) {
+    init(transitions: [From]) {
         self.transitions = transitions
-    }
+        let transitionsForStates = self.transitions.reduce(into: [AnyHashable: (State) -> [AnyHashable: (Event) -> State?]]()) { accumulator, from in
+            let existingTranstionsForState = accumulator[from.id]
+            accumulator[from.id] = { state in from.computeTransitionsForEvents(for: state, existingTranstionsForState: existingTranstionsForState) }
+        }
+        self.reducer = { state, event -> State in
+            if
+                let transitionsForState = transitionsForStates[state.instanceId],
+                let transitionForEvent = transitionsForState(state)[event.instanceId],
+                let newState = transitionForEvent(event) { return newState }
 
-    public var entries: [TransitionId: (State, Event) -> State] {
-        self.transitions.reduce(into: [TransitionId: (State, Event) -> State]()) { accumulator, transition in
-            accumulator.merge(transition.entries, uniquingKeysWith: { $1 })
+            if
+                let transitionsForState = transitionsForStates[state.instanceId],
+                let transitionForEvent = transitionsForState(state)[AnyEvent.id],
+                let newState = transitionForEvent(event) { return newState }
+
+            if
+                let transitionsForState = transitionsForStates[AnyState.id],
+                let transitionForEvent = transitionsForState(state)[event.instanceId],
+                let newState = transitionForEvent(event) { return newState }
+
+            if
+                let transitionsForState = transitionsForStates[AnyState.id],
+                let transitionForEvent = transitionsForState(state)[AnyEvent.id],
+                let newState = transitionForEvent(event) { return newState }
+
+            return state
         }
     }
 }
 
 // MARK: modifiers
 public extension Transitions {
+    /// Disables all the transitions of the state machine, as long as the `disabled` condition is true
+    /// - Parameter disabled: the condition that disables the transitions
+    /// - Returns: the  transitions
     func disable(_ disabled: @escaping () -> Bool) -> Self {
-        let disabledTransitions = self.transitions.map { $0.disable(disabled) }
-
-        return Transitions(transitions: disabledTransitions)
-    }
-}
-
-public extension Transitions {
-    var reducer: (State, Event) -> State {
-        let entries = self.entries
-
-        return { state, event -> State in
-            let transitionId = TransitionId(stateId: state.instanceId, eventId: event.instanceId)
-
-            if let reducer = entries[transitionId] {
-                return reducer(state, event)
-            }
-
-            let transitionAnyEventId = TransitionId(stateId: state.instanceId, eventId: AnyEvent.id)
-            if let reducer = entries[transitionAnyEventId] {
-                return reducer(state, AnyEvent())
-            }
-
-            let transitionAnyStateId = TransitionId(stateId: AnyState.id, eventId: event.instanceId)
-            if let reducer = entries[transitionAnyStateId] {
-                return reducer(AnyState(), event)
-            }
-
-            let transitionAnyId = TransitionId(stateId: AnyState.id, eventId: AnyEvent.id)
-            if let reducer = entries[transitionAnyId] {
-                return reducer(AnyState(), AnyEvent())
-            }
-
-            return state
-        }
+        Transitions(transitions: self.transitions.map { $0.disable(disabled) })
     }
 }