+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 **v0.3.0 - Tyranus**:
 
 - Feedback: introduce the "on:" keyword to explicitly declare the type of state that concerns the side effect
+- Feedback: replace the parameter "sideEffect" by "perform" to have a nice readable sentence: ...(on: Loading.self, ..., perform: sideEffect)
+- State Machine: introduce a new DSL based on From/On that allows to group transitions from the same state type
+- State Machine: provide assert functions to ease the unit tests of transitions
 
 **v0.2.0 - Vader**: